+++ assets/src/apps/v2/gl_eink.tsx
@@ -40,7 +40,7 @@ import BottomScreenFiller from "Components/v2/eink/bottom_screen_filler";
 import OvernightDepartures from "Components/v2/eink/overnight_departures";
 import MultiScreenPage from "Components/v2/multi_screen_page";
 import SimulationScreenPage from "Components/v2/simulation_screen_page";
-import ExceptionCatcher from "./exception_catcher";
+import ExceptionCatcher from "Components/v2/exception_catcher";
 
 const TYPE_TO_COMPONENT = {
   screen_normal: NormalScreen,
